// src/app/services/api/mtc/model/assistitoRequest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface AssistitoRequest { 
    /**
     * Cognome
     */
    cognome: string;
    /**
     * Data di nascita (yyyy-mm-dd)
     */
    dataNascita: string;
    /**
     * Email
     */
    email: string;
    /**
     * ID assistito
     */
    id: string;
    /**
     * Nome
     */
    nome: string;
}