// src/app/services/api/mtc/model/createRichiestaTelevisitaRequest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AssistitoRequest } from './assistitoRequest';
import { MedicoRequest } from './medicoRequest';
import { PartecipanteRequest } from './partecipanteRequest';


export interface CreateRichiestaTelevisitaRequest { 
    /**
     * Altri partecipanti della televisita
     */
    altriPartecipanti?: Array<PartecipanteRequest>;
    /**
     * Assistito
     */
    assistito: AssistitoRequest;
    /**
     * ID cartella
     */
    idCartella?: string;
    /**
     * Medico
     */
    medico: MedicoRequest;
    /**
     * Campo motivazione
     */
    motivazione: string;
    /**
     * Campo note
     */
    note?: string;
}