// src/app/services/api/mtc/model/findAllDurataResponse.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DurataResponse } from './durataResponse';


export interface FindAllDurataResponse { 
    durate?: Array<DurataResponse>;
}