// src/app/services/api/mtc/model/medicoRequest.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface MedicoRequest { 
    /**
     * Codice
     */
    codice: string;
    /**
     * Cognome
     */
    cognome: string;
    /**
     * Email
     */
    email: string;
    /**
     * ID medico
     */
    id: string;
    /**
     * Nome
     */
    nome: string;
}